pyproject.toml
[project]
name = "kanae"
version = "0.1.0"
description = "Internal backend server for ACM @ UC Merced"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiocache~=0.12.3",
    "aiohttp[speedups]~=3.13.5",
    "argon2-cffi~=25.1.0",
    "asyncpg~=0.31.0",
    "blake3~=1.0.8",
    "fastapi~=0.136.0",
    "fastapi-pagination~=0.15.12",
    "gunicorn~=25.3.0",
    "httptools~=0.7.1",
    "limits~=5.8.0",
    "orjson~=3.11.8",
    "prometheus-fastapi-instrumentator~=7.1.0",
    "pybase62~=1.0.0",
    "python-dateutil~=2.9.0.post0",
    "pyyaml~=6.0.3",
    "uvicorn~=0.46.0",
    "uvloop~=0.22.1 ; sys_platform != 'win32'",
    "valkey-glide~=2.3.1",
    "winloop~=0.6.2 ; sys_platform == 'win32'",
]

[dependency-groups]
dev = [
    "colorama~=0.4.6 ; sys_platform == 'win32'",
    "watchfiles~=1.1.1",
    { include-group = "lint" },
    { include-group = "test" },
]
lint = [
    "ruff~=0.15.11",
    "ty~=0.0.31",
]
test = [
    "asgi-lifespan~=2.1.0",
    "hiro~=1.1.1",
    "httpx~=0.28.1",
    "pytest~=9.0.3",
    "pytest-asyncio~=1.3.0",
    "testcontainers~=4.14.2",
]

[tool.ty.src]
include = ["server"]
exclude = [
    "__pycache__", 
    "docs",
    "docker",
    "tests"
]

[tool.ruff]
line-length = 88
extend-include = [
    "core.py",
    "launcher.py",
    "utils/**"
]
extend-exclude = ["**/__pycache__", "tests"]

[tool.ruff.lint]
select = [
    "A", "ANN", "ASYNC", "B", "BLE", "C4", "COM", "DTZ", "E",
    "EM", "ERA", "F", "FA", "FAST", "FBT", "FLY", "FURB", "G", "I", "ICN", "ISC", "LOG", "NPY",
    "PD", "PERF", "PGH", "PIE", "PLC", "PLE", "PLR", "PLW", "PTH", "PYI",
    "Q", "Q003", "RET", "RSE", "RUF", "S", "SIM", "SLOT", "T20", "TC", "TID",
    "TRY", "UP", "W", "YTT"
]
ignore = [
    "B904",     # Niche standard
    "E501",     # Ruff suggested format
    "FBT003",   # Inherently incorrect to enforce this 
    "PLR0913",  # Violations of these are not related to bad code
    "PLR2004",  # Magic numbers are fine sometimes
    "SIM105",   # Prefer not to use due to stylistic choices
    "UP045",    # Explicitly use typing.Optional to not confuse between union operators.
                # See https://discuss.python.org/t/clarification-for-pep-604-is-foo-int-none-to-replace-all-use-of-foo-optional-int/26945/19
    
    # These are recommended by Ruff if the formatter is to be used: https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
    "W191",
    "E111",
    "E114",
    "E117",
    "D206",
    "D300",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "ISC001",
    "ISC002"
]
fixable = ["ALL"]
unfixable = [
    "E501",   # Ruff suggests their own formatting
    "ERA"     # Leave commented-out code for reference purposes
]
[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
minversion = "9.0"
addopts = "-ra"
testpaths = [
    "tests"
]
pythonpath = ["server"]
asyncio_default_fixture_loop_scope = "function"
